MisterTube (New Re:Branding)
This Is inspired for by the use of uBlock Origin And Violent Money Script Extensions, But for YouTube.
It's Not even at beta stage. It will be an awesome Extension. Which I will prove that by making massive changes in my 1-month leave. 
I was trying to make it in Manifest v2 Because I didn't know Chrome based Browser didn't allow any extension of v2.